requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.21
python-dotenv==1.0.0
pydantic==2.4.2
requests==2.31.0
alembic==1.12.1
supabase==2.28.0